fix(migrations): replace numeretic to integer

+++ migrations/000001_init.sql
@@ -2,7 +2,7 @@
 CREATE TABLE IF NOT EXISTS subscriptions (
     id BIGSERIAL PRIMARY KEY,
     service_name VARCHAR(255) NOT NULL,
-    price numeric(12, 2) NOT NULL,
+    price INTEGER NOT NULL,
     user_id UUID NOT NULL,
     start_date DATE NOT NULL,
     end_date DATE NOT NULL,
